Move PDF export modal out of shared folder and into the only folder where it is even used
