+++ front/app/components/admin/FormSync/PDFExportModal/MultilocFieldCollapsible.tsx
